kegwasher-service/requirements.txt
Adafruit-CharLCD==1.1.1
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.5
RPi.GPIO==0.7.0
smbus2==0.3.0
spidev==3.4